tools/tts-en/requirements.txt
coqui-tts==0.27.2

